--- a/Regress/Scripts/new_timings.py
+++ b/Regress/Scripts/new_timings.py
@@ -3,34 +3,40 @@ import time, gzip, random, os, sys
 from rdkit import Chem
 from rdkit.Chem import AllChem
 from rdkit.RDLogger import logger
+
+dname = os.path.dirname(__file__)
+def data(fname):
+  return os.path.join(dname, '..', 'Data', fname)
+
 logger = logger()
 
 tests = [1] * 1001
 if len(sys.argv) > 1:
   tests = [0] * 1001
-  tests[1] = 1
   for x in sys.argv[1:]:
     x = int(x)
     tests[x] = 1
 ts = []
 mols = []
-lines = gzip.open('../Data/znp.50k.smi.gz', 'rt').readlines()
-logger.info('mols from smiles')
-nMols = 0
-nBad = 0
-t1 = time.time()
-for line in lines:
-  line = line.strip().split(' ')
-  m = Chem.MolFromSmiles(line[0])
-  if m:
-    nMols += 1
-    mols.append(m)
-  else:
-    nBad += 1
 
-t2 = time.time()
-logger.info('Results1: %.2f seconds, %d passed, %d failed' % (t2 - t1, nMols, nBad))
-ts.append(t2 - t1)
+if tests[0]:
+  lines = gzip.open(data('znp.50k.smi.gz'), 'rt').readlines()
+  logger.info('mols from smiles')
+  nMols = 0
+  nBad = 0
+  t1 = time.time()
+  for line in lines:
+    line = line.strip().split(' ')
+    m = Chem.MolFromSmiles(line[0])
+    if m:
+      nMols += 1
+      mols.append(m)
+    else:
+      nBad += 1
+
+  t2 = time.time()
+  logger.info('Results1: %.2f seconds, %d passed, %d failed' % (t2 - t1, nMols, nBad))
+  ts.append(t2 - t1)
 
 if tests[1]:
   logger.info('Writing: Canonical SMILES')
@@ -42,7 +48,7 @@ if tests[1]:
   ts.append(t2 - t1)
 
 if tests[2]:
-  sdData = gzip.open('../Data/mols.1000.sdf.gz', 'rb').read()
+  sdData = gzip.open(data('mols.1000.sdf.gz'), 'rb').read()
   logger.info('mols from sdf')
   suppl = Chem.SDMolSupplier()
   suppl.SetData(sdData)
@@ -61,7 +67,7 @@ if tests[2]:
   ts.append(t2 - t1)
 
 if tests[3] or tests[4] or tests[5]:
-  pattData = gzip.open('../Data/queries.txt.gz', 'rt').readlines()
+  pattData = gzip.open(data('queries.txt.gz'), 'rt').readlines()
   pattData = [x.strip().replace('[H]', '').replace('()', '') for x in pattData]
   logger.info('patterns from smiles')
   patts = []
@@ -104,7 +110,7 @@ if tests[6] or tests[7] or tests[8]:
   logger.info('reading SMARTS')
   patts = []
   t1 = time.time()
-  for line in open('../Data/RLewis_smarts.txt'):
+  for line in open(data('RLewis_smarts.txt')):
     line = line.strip()
     if line == '' or line[0] == '#':
       continue