This has the following two performance improvements to findSSSR that leads to speeding up Chem.SanitizeMol substantially on large systems  (thousands of atoms):

- When the fragment does not contain any rings as determined by Euler's formula (E-V+1 = #rings) such as a water box, skip the SSSR detection entirely. ~1800x faster on 9,000 waters with explicit hydrogens. 
- Avoid an O(N^2) loop to determine the number of bonds in the fragment. Use a cheaper method of summing the atom degrees and dividing by two. (There was already a loop over atom degrees, so easy to highjack that loop to do this operation). ~10x on 60,000 carbon atom chain with a cyclic-propane attached. 

The result is no real measurable effect on the new_timings.py smiles parsing unfortunately, but didn't make it worse :-) 
before: new_timings.py 0

[11:51:50] INFO: mols from smiles
[11:52:11] INFO: Results1: 21.40 seconds, 50000 passed, 0 failed
times:  21.4
[11:52:14] INFO: mols from smiles
[11:52:36] INFO: Results1: 21.62 seconds, 50000 passed, 0 failed
times:  21.6
[11:52:39] INFO: mols from smiles
[11:53:00] INFO: Results1: 21.48 seconds, 50000 passed, 0 failed
times:  21.5

after: new_timings.py 0

[11:58:27] INFO: mols from smiles
[11:58:49] INFO: Results1: 21.60 seconds, 50000 passed, 0 failed
times:  21.6
[11:58:52] INFO: mols from smiles
[11:59:13] INFO: Results1: 21.25 seconds, 50000 passed, 0 failed
times:  21.3
[11:59:16] INFO: mols from smiles
[11:59:37] INFO: Results1: 21.36 seconds, 50000 passed, 0 failed
times:  21.4

Attaching the timings and scripts used for them to the pull request. Doesn't appear to be a need to check them into the code since they are engineered corner cases for the extreme (though I don't think a water box as extreme :-). 

[time_sssr_on_water.txt](https://github.com/rdkit/rdkit/files/558400/time_sssr_on_water.txt)
[time_sssr_on_long_chain_with_some_rings.txt](https://github.com/rdkit/rdkit/files/558402/time_sssr_on_long_chain_with_some_rings.txt)
[time_sssr_on_long_chain_with_one_ring.txt](https://github.com/rdkit/rdkit/files/558404/time_sssr_on_long_chain_with_one_ring.txt)
[sssr_opt_tweak.txt](https://github.com/rdkit/rdkit/files/558407/sssr_opt_tweak.txt)